# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/opening": {
            "post": {
                "description": "Creates a new opening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openings"
                ],
                "summary": "Create Opening",
                "parameters": [
                    {
                        "description": "Data needed to create an opening",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOpening"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestResponse"
                        }
                    }
                }
            }
        },
        "/opening/{id}": {
            "get": {
                "description": "Gets an opening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openings"
                ],
                "summary": "Get Opening",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of the opening to get",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a opening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openings"
                ],
                "summary": "Update Opening",
                "parameters": [
                    {
                        "description": "Data needed to update an opening",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateOpening"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Id of the opening to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an opening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openings"
                ],
                "summary": "Delete Opening",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of the opening to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestResponse"
                        }
                    }
                }
            }
        },
        "/openings": {
            "get": {
                "description": "Gets all openings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openings"
                ],
                "summary": "Get All Openings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateOpening": {
            "type": "object",
            "required": [
                "company",
                "link",
                "location",
                "remote",
                "role",
                "salary"
            ],
            "properties": {
                "company": {
                    "type": "string",
                    "minLength": 10
                },
                "link": {
                    "type": "string"
                },
                "location": {
                    "type": "string",
                    "minLength": 10
                },
                "remote": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string",
                    "minLength": 10
                },
                "salary": {
                    "type": "integer"
                }
            }
        },
        "dto.RequestResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateOpening": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string",
                    "minLength": 10
                },
                "link": {
                    "type": "string"
                },
                "location": {
                    "type": "string",
                    "minLength": 10
                },
                "remote": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string",
                    "minLength": 10
                },
                "salary": {
                    "type": "integer"
                }
            }
        }
    }
}